Make C-t in bash/zsh respect FZF_DEFAULT_COMMAND.
I had to move the configuration into the generated config files instead
of where one would normally put it (e.g. ~/.bashrc). This is to properly
support tmux, since when using tmux a new shell is spawned that may not
be configured the same way.

I was unable to figure out how to make this work for fish. For some
reason the new shell in the tmux pane does not get variables set with
`set -x` (I tried placing them as the first lines of `fzf_key_bindings`).