shell/key-bindings.bash: Correct the keybindings for search direction
The readline man page [1], and the bash man page [2], state that C-r and C-s
are to be used to search backwards and forwards in an incremental search.

This patch removes the search sort toggle from fzf, and adds bindings for
forward and backwards search as according to the readline bindings to bring
compatability closer.

[1] http://linux.die.net/man/3/readline
[2] http://linux.die.net/man/1/bash