shell/key-bindings.bash: Add support for CTRL-G to reset search
As per the readline man page [1], and the bash man page [2], while in an
incremental search, the keyseqence C-g should abort the search, and
restore the original line.

This patch changes the __fzf_history__ function to change the READLINE_LINE
variable through a bind -x command, instead of doing shell expansion on
$(__fzf_history__) as before. This allows us to use the fzf --accept argument
to catch a ctrl-g press, and just not change the READLINE_LINE variable, thus
keeping the line in it's original state.

An additional change is to save doing shell expansion, and instead use the
`history -p` syntax, and only fall back to shell history expansion if that
fails. In testing, I found this to only fail if you are recalling the last
entered command in a non-fresh session (ie, a command has been typed, and you
search and replay that last command). In this case, the fallback will first
test whether the shell can do history expanion, and if it can, sends back a
keysequence to expand and go to end of line. If it can't, a grep is done on
the output of the history, which currently breaks when the last command is a
multiple line command.

The C-r keybinding is actually bound to C-xC-h which is a new binding (-x) to
call the new __fzf_history__ function, and then also does C-xC-f, which is
either bound to nothing, or bound to a "finish" sequence. In this patch, the
only "finish" sequence is history expansion and end of line. In the future,
this could be anything.

[1] http://linux.die.net/man/3/readline
[2] http://linux.die.net/man/1/bash