shell/key-bindings.bash: Add ctrl-o mode when searching history
From the bash man page [1], C-o is bound to operate-and-get-next (when in
emacs mode only). This command will run the currently selected command, then
get the next command from the history to display (and potentially edit). If
repeated, then it goes on until it reaches the end of the history, then jumps
back the number of commands run. This allows you to cycle through a set number
of commands in an ordering.

This patch adds this functionality to the ctrl-r search history. Now when you
are in emacs mode, and you select a history element, and type ctrl-o, it will
run that command, and retrieve the next one. If you then repeat pressing ctrl-o
it will run all the commands up until the end of history, then cycle. How this
is done technically is by storing an offset back from the end of history, as
new commands will append to the history, so that offset will always be correct
and it will pick up any modifications you made.

Currently it doesn't support up arrow (patch to come).

Also overridden is C-m (or accept-line), in order to check whether we are
ending a ctrl-o cycle, or just accepting the line in a ctrl-o. This is done via
a variable `_fzf_history_ctrl_o_present` which is toggled on when in ctrl-o
mode and off when a line is accepted. If it is off when the line is accepted,
then enter was pushed by the user, so we reset the counter offset.

[1] http://linux.die.net/man/1/bash