shell/key-bindings.bash: Add previous and next history commands for ctrl-o
This patch adds the previous and next history commands (up and down arrow) to
the ctrl-o mode. These should increment and decrement the offset until the mode
is entered, then it should not change until the mode is exitted. This is done
with a new variable `_fzf_history_scrollable`, which is set at start, and when
ctrl-o mode exits (tested in `__zfz_accept_line__`), and unset in when ctrl-o
mode is entered (`__zfz_ctrl_o__`).