Prefer tmux for launching in Vim
There is currently a [bug in Neovim][1] where opening a terminal buffer
by exec'ing a command that has a pipe in it causes the buffer to be
blank until `<ESC>` is pressed. While not strictly an issue with fzf, it
affects users like me who use a custom `FZF_DEFAULT_COMMAND`.

This change makes fzf prefer tmux over Neovim for launching a fzf split.
It's reasonable to assume that if a Neovim user under tmux either
- Doesn't care about the type of split, or
- Prefers tmux panes to nvim embedded terminals, as panes the main
  purpose of tmux.

Thus, this change works around the Neovim issue in a non-temporary,
non-hacky way; looks like a win-win.

[1]: https://github.com/neovim/neovim/issues/4487