Overhaul fish functions
Replace the "temp file" workaround with the "read" function: it's
simpler and faster.

Use proper escaping, remove the custom function.

The "file" widget uses last token as root for the "find" command.
This replaces the equivalent of '**' completion in bash/zsh.
The "$cwd" non-expanded variable can be used in FZF_CTRL_T_COMMAND to
set the root.