Append '/' to folders in the file widget
This helps differentiating folders from files.

Besdies, writing a folder with an appended slash on the commandline is
usually what the user wants: the user can keep on completing/fzf-ing, it
saves one key stroke.

The `+` argument makes sure `find` will spawn only one `printf` process
for all folders. Because of this, the folders will be printed at the
end, root included. Therefore, the `sed` command would not discard the
root properly: we use `-mindepth ` instead (which saves one process by
the way).