[vim] Recreate command and old public function names
All function definitions are moved into the autoload namespace fzf#core.
The previously defined and publicly visible functions fzf#shellescape,
fzf#run and fzf#wrap still exist for compatibility and just wrap the
corresponding function in fzf#core.

The main purpose of this is to reduce startup time.  Most of the code of
this plugin will now only be loaded when it is called.

The functions where not simply moved to the fzf autoload namespace
because in this case the loading of the plugin/fzf.vim file from the
junegunn/fzf.vim repository would have triggered the autoloading
mechanism for them.  This would have defeated the purpose mentioned
above.