Do not match HostKeyAlgorithms in SSH config file
Modern versions of SSH support a config option called HostKeyAlgorithms
that identifies the algorithms that will be offered or accepted (in the
case of SSHD). Since _fzf_complete_ssh does a case insensitive grep for
host, this line--if present--is also picked up.

The line being changed is:
https://github.com/junegunn/fzf/blob/master/shell/completion.bash#L244

The addition is: command grep -v 'HostKeyAlgorithms |'