Two fixes and one cosmetics for install script
Fix: line 94: it is possible for `fzf --version` to fail with output
that satisfies awk, therefore $? is 0 in spite of the error. Fixed by
adding pipefail. Seen in attached install log. Due to previous error in
the same log.

Change: line 113 (cosmetics). Reverse `->` link notation to match that
of `ls -l`.

Fix: line 114: It is possible for the bin directory to not exist. Fixed
by adding `mkdir -p`. Seen in attached install log. Execution continued
to expose the error on line 94 (fixed above).

Attached: install log file