shell: add support to deduplicate history command output
This change allows to deduplicate commands from the shell history
when filtering with fzf.

To enable deduplication, set "FZF_HIST_FIND_NO_DUPS" to "on":

    export FZF_HIST_FIND_NO_DUPS=on

On zsh, fall back to reading the "HIST_FIND_NO_DUPS" shell option
which can be set as follows:

    setopt HIST_FIND_NO_DUPS

A big »Thanks« goes to Felix Breuer for his idea which this change is
based upon ([0]).

Updates [1] and [2].

[0]: https://github.com/junegunn/fzf/pull/1363#issuecomment-598228199
[1]: https://github.com/junegunn/fzf/pull/1287
[2]: https://github.com/junegunn/fzf/pull/1363