Allow inserting multiple lines from history
Allows inserting multiple lines from the history at once.

Preserves the existing behaviour, e.g. moving the cursor to the end of
the input line.

Newline-separates the lines if multiple are specified. I didn't add a
trailing newline to keep compatibility with the existing implementation.

Confirmed this works with `\t` and `\n` chars, and assuming `echo`,
being a zsh builtin, will handle the unescaping adequately.

Fixes: https://github.com/junegunn/fzf/issues/1806
Refs: https://superuser.com/questions/1564346/insert-multiple-lines-from-zsh-history