Move shell string quoting to platform-specific utilities
Move the quoting of shell strings into the platform-specific files
where the rest of shell command execution code is already, instead
of manually checking which quoting to use in terminal.go. This is
a bit cleaner, and makes it easier to support different shell commands.
The actual quoting implementation is still in platform-agnostic util.go
so it can be tested irregardless of the platform.