Enable alt+c bash binding with version > 4 syntax.
This patch introduces the same backwards compatible logic as the other
bash bindings, I believe the logic for alt+c was perhaps just missed in
a previous patch to use the bash > 4 syntax, which is cleaner and
actually resolves an issue with the terminal bash prompt overlapping in
a strange way.

This is mostly a cosmetic fix but also fixes another issue with the fzf
command exiting cleanly when aborted.