Change the version check invocation in install.ps1
When install.ps1 is executed inside MSYS (or cygwin I suspect) cmd
is a shell script that invokes the real cmd.  But in install.ps1 it
tries to open this shell script instead of executing it.

AFAICT the only reason to use cmd /c is to capture stderr because
there is a bug with powersehll; but fzf prints the version information
on stdout so we don't need to care about stderr.  So we can just call
the binary directly.  This avoids the cmd issues.