[zsh]: bring keybind functionalities on par with Fish
Prior to this commit, the zsh keybindings did not support a prefix
search and did not support using an initial directory for commands.

For instance, the following operation was impossible prior to this
commit.
Pressing `Ctrl-T` while the cursor was after a directory would
not start the search from the given directory.
i.e. `> ls src/_` (where underscore denotes the current cursor),
the search for files would always start in '.'.

With this commit, zsh's behavior is now inline with Fish's behavior
and will correctly start searching in the first existing directory.

The `M-c` (alt-c) behavior was also fixed to be compatible with Fish's
behavior.