Check if SHELL supports FZF_DEFAULT_COMMAND
When FZF_DEFAULT_COMMAND is unset, it defaults to a command that depends
on 'set -o pipefail'. Since not every shell supports this option, bash
was hardcoded as the default shell to be used for this specific case.
With this commit, other shells that support this option are first
compared to SHELL before depending/falling back to bash.