fix: support comments in FZF_DEFAULT_OPTS_FILE by sanitizing content
Fixes: #3961

Added a sanitization step to the `FZF_DEFAULT_OPTS_FILE` processing,
allowing users to include comments in their configuration.

The sanitizer skips empty lines and lines starting with a `#`, inserting
the remaining options into a buffer that will be passed to the parser.

Inline comments are not supported. Additionally, any argument that
includes a `#` should remain unaffected by this change.